feat(key on repo url): consolidate validation for repo creation to reduce duplication
